requirements.txt
# Core libraries for the RAG pipeline
supabase
sentence-transformers
langchain

# For running the notebook environment
jupyter

# For managing environment variables
python-dotenv

# extract text from PDFs
python-multipart
pymupdf